requirements.txt
streamlit
langchain
langchain_core
langchain_community
langchain_openai
openai
python-dotenv
